Move burndown utilities to modes
