+++ python/labours/modes/ownership.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Tuple
 import numpy
 
 from labours.plotting import apply_plot_style, deploy_plot, get_plot_path, import_pyplot
-from labours.utils import default_json, floor_datetime, parse_date, import_pandas
+from labours.utils import default_json, floor_datetime, import_pandas, parse_date
 
 
 def load_ownership(
